+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/ContentBlockingController.java
@@ -318,15 +318,6 @@ public class ContentBlockingController {
         /**
          * Indicates that content that would have been blocked has instead been
          * replaced with a shim.
-         * @deprecated use {@link #REPLACED_TRACKING_CONTENT} instead.
-         */
-        @Deprecated
-        @DeprecationSchedule(version = 86, id = "unsafe-content")
-        public static final int REPLACED_UNSAFE_CONTENT        = 0x00000010;
-
-        /**
-         * Indicates that content that would have been blocked has instead been
-         * replaced with a shim.
          */
         public static final int REPLACED_TRACKING_CONTENT       = 0x00000010;
 
@@ -352,7 +343,7 @@ public class ContentBlockingController {
                       Event.COOKIES_BLOCKED_TRACKER, Event.COOKIES_BLOCKED_SOCIALTRACKER,
                       Event.COOKIES_BLOCKED_ALL, Event.COOKIES_PARTITIONED_FOREIGN,
                       Event.COOKIES_BLOCKED_FOREIGN, Event.BLOCKED_SOCIALTRACKING_CONTENT,
-                      Event.LOADED_SOCIALTRACKING_CONTENT, Event.REPLACED_UNSAFE_CONTENT })
+                      Event.LOADED_SOCIALTRACKING_CONTENT, Event.REPLACED_TRACKING_CONTENT })
             /* package */ @interface LogEvent {}
 
             /**